refactor doc package
