+++ citizen-intelligence-agency/src/main/java/com/hack23/cia/web/impl/ui/application/views/common/pagemode/CommonViewConstants.java
@@ -4,18 +4,18 @@ package com.hack23.cia.web.impl.ui.application.views.common.pagemode;
  * Constants for common views and pages.
  */
 public interface CommonViewConstants {
-    
+
     /** Dashboard constants */
     String DASHBOARD_TITLE = "CitizenIntelligence Agency::Dashboard Overview";
     String DASHBOARD_HEADER = "Dashboard Overview";
     String DASHBOARD_DESCRIPTION = "Visualize political activity in Sweden, present key performance indicators and metadata.";
-    
+
     /** Dashboard monarch section constants */
     String MONARCH_SECTION_TITLE = "Monarch";
     String MONARCH_CURRENT_INFO = "Head of state(King): Carl Gustaf Folke Hubertus since 15 September 1973";
     String MONARCH_SUCCESSOR_INFO = "Future head of state(Queen): Victoria Ingrid Alice Désirée";
     String MONARCH_COURT_TITLE = "Kungliga hov- och slottsstaten(The Royal Court)";
-    
+
     /** Financial section constants */
     String FINANCIAL_SECTION_HEADER = "Financial";
     String FINANCIAL_STAT_INCOME = "Income(B SEK)";
@@ -27,7 +27,7 @@ public interface CommonViewConstants {
     String FINANCIAL_STAT_INCOME_B = "Income(B SEK)";
     String FINANCIAL_STAT_SPENDING_B = "Spending(B SEK)";
     String FINANCIAL_STAT_RESULT_B = "Result(B SEK)";
-    
+
     /** Government section constants */
     String GOV_SECTION_TITLE = "Government";
     String GOV_STAT_MEMBERS = "Members";
@@ -36,55 +36,55 @@ public interface CommonViewConstants {
     String GOV_STAT_BODIES = "Government bodies";
     String GOV_STAT_HEADCOUNT = "Headcount";
     String GOV_OFFICES_TITLE = "Regeringskansliet(Government Offices)";
-    
+
     /** Parliament section constants */
     String PARLIAMENT_SECTION_TITLE = "Parliament";
     String PARLIAMENT_ADMIN_TITLE = "Riksdagsförvaltningen(the Riksdag administration)";
-    
+
     /** Risk section constants */
     String RISK_BY_TYPE_TITLE = "Number of risk by each type";
     String RISK_BY_SEVERITY_TITLE = "Number of risk by severity";
-    
+
     /** The overview title header. */
-    String OVERVIEW_TITLE_HEADER = "Overview"; 
-    
+    String OVERVIEW_TITLE_HEADER = "Overview";
+
     /** The overview title. */
     String OVERVIEW_TITLE = "Overview";
-    
+
     /** The overview description. */
     String OVERVIEW_DESCRIPTION = "Visualize political activity in Sweden, present key performance indicators and metadata.";
-    
+
     /** The visit history title header. */
     String VISIT_HISTORY_TITLE_HEADER = "Page Visits";
-    
+
     /** The visit history title. */
     String VISIT_HISTORY_TITLE = "Page Visits";
-    
+
     /** The visit history description. */
     String VISIT_HISTORY_DESCRIPTION = "History of page visits to understand user engagement.";
-    
+
     /** The login title header. */
     String LOGIN_TITLE_HEADER = "Login";
-    
+
     /** The login title. */
-    String LOGIN_TITLE = "Login User"; 
-    
+    String LOGIN_TITLE = "Login User";
+
     /** The login description. */
     String LOGIN_DESCRIPTION = "Secure login for agency users.";
-    
+
     /** The register title header. */
     String REGISTER_TITLE_HEADER = "Register";
-    
-    /** The register title. */  
+
+    /** The register title. */
     String REGISTER_TITLE = "Register User";
-    
+
     /** The register description. */
     String REGISTER_DESCRIPTION = "Register new users with essential details.";
-    
+
     /** Chart constants */
     String CHART_CURRENT_COMMITTEES = "Current Committees";
     String CHART_CURRENT_PARTIES = "Current Parties, total political days served";
-    
+
     /** Form constants */
     String LOGIN_FORM_TITLE = "Login";
     String LOGIN_USER_FORM_TITLE = "Login user";
