--- citizen-intelligence-agency/src/main/java/com/hack23/cia/web/impl/ui/application/views/user/document/pagemode/DocumentsOverviewPageModeConstants.java
@@ -1,37 +1,0 @@
-package com.hack23.cia.web.impl.ui.application.views.user.document.pagemode;
-
-/**
- * The Interface DocumentsOverviewPageModeConstants.
- */
-public interface DocumentsOverviewPageModeConstants {
-
-    /** The document type text. */
-    String DOCUMENT_TYPE_TEXT = "Document Type";
-
-    /** The document status text. */
-    String DOCUMENT_STATUS_TEXT = "Document Status";
-
-    /** The document size text. */
-    String DOCUMENT_SIZE_TEXT = "Size";
-
-    /** The document chart colors. */
-    String[] DOCUMENT_CHART_COLORS = { "#FF9900", "#109618", "#990099", "#3366CC", "#DC3912" };
-
-    /** The document activity chart description. */
-    String DOCUMENT_ACTIVITY_CHART_DESCRIPTION = "Document activity over time";
-
-    /** The document type chart description. */
-    String DOCUMENT_TYPE_CHART_DESCRIPTION = "Document type distribution";
-
-    /** The document status chart description. */
-    String DOCUMENT_STATUS_CHART_DESCRIPTION = "Document status distribution";
-
-    /** The document activity heading. */
-    String DOCUMENT_ACTIVITY_HEADING = "Document activity";
-
-    /** The document type distribution heading. */
-    String DOCUMENT_TYPE_DISTRIBUTION_HEADING = "Document type distribution";
-
-    /** The document status distribution heading. */
-    String DOCUMENT_STATUS_DISTRIBUTION_HEADING = "Document status distribution";
-}
