+++ citizen-intelligence-agency/src/main/java/com/hack23/cia/web/impl/ui/application/views/user/documents/DocumentsView.java
@@ -16,7 +16,7 @@
  *	$Id$
  *  $HeadURL$
 */
-package com.hack23.cia.web.impl.ui.application.views.user.document;
+package com.hack23.cia.web.impl.ui.application.views.user.documents;
 
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.ApplicationContext;
@@ -29,26 +29,26 @@ import com.hack23.cia.web.impl.ui.application.views.user.common.AbstractUserView
 import com.vaadin.spring.annotation.SpringView;
 
 /**
- * The Class PartyView.
+ * The Class DocumentsView.
  */
 @Service
 @Scope(value = ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-@SpringView(name = SearchDocumentView.NAME)
-public final class SearchDocumentView extends AbstractUserView {
+@SpringView(name = DocumentsView.NAME)
+public final class DocumentsView extends AbstractUserView {
 
 	/** The Constant NAME. */
-	public static final String NAME = UserViews.SEARCH_DOCUMENT_VIEW_NAME;
+	public static final String NAME = UserViews.DOCUMENTS_VIEW_NAME;
 
 	/** The Constant serialVersionUID. */
 	private static final long serialVersionUID = 1L;
 
 	/**
-	 * Instantiates a new search document view.
+	 * Instantiates a new document view.
 	 *
 	 * @param context
 	 *            the context
 	 */
-	public SearchDocumentView(final ApplicationContext context) {
+	public DocumentsView(final ApplicationContext context) {
 		super(context.getBeansOfType(PageModeContentFactory.class), NAME);
 
 	}
